+++ public/js/ng/datasource-edit/datasource-edit.module.js
@@ -1,5 +1,0 @@
-(function () {
-    'use strict';
-
-    angular.module('app.datasource-edit', ['app.core']);
-})();