+++ public/js/ng/datasource-list/datasource-list.module.js
@@ -1,5 +1,0 @@
-(function () {
-    'use strict';
-
-    angular.module('app.datasource-list', ['app.core', 'app.table']);
-})();